requirements.txt
# Deep Learning Framework
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter
torch-sparse
torch-cluster

# Model Serialization
safetensors>=0.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# OCR and PDF Processing
paddleocr>=2.7.0
paddlepaddle-gpu
PyMuPDF>=1.23.0
Pillow>=10.0.0

# Web Scraping
selenium>=4.15.0
webdriver-manager>=4.0.0
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Async Processing
asyncio
aiofiles>=23.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Utilities
tqdm>=4.66.0
tensorboard>=2.14.0
python-json-logger>=2.0.7

# Optional: For API access (if using extraction features)
# openai>=1.0.0

# Development Tools (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
